requirements.txt
streamlit
langchain
openai
faiss-cpu
tiktoken
python-dotenv
transformers
sentence-transformers
pymupdf
pdfplumber
spacy
